feat: add search limit
